requirements.txt
asgiref==3.6.0
Django==4.1.8
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
PyJWT==2.6.0
pymongo==4.3.3
pytz==2023.3
sqlparse==0.2.4
